Python cleanup
